+++ include/sqlpp11/mysql/bind_result.h
@@ -26,91 +26,21 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <iostream>
-#include <memory>
-#include <vector>
 #include <sqlpp11/chrono.h>
 #include <sqlpp11/exception.h>
+#include <sqlpp11/mysql/detail/prepared_statement_handle.h>
 #include <sqlpp11/mysql/sqlpp_mysql.h>
 
+#include <iostream>
 #ifdef _MSC_VER
 #include <iso646.h>
 #endif
+#include <memory>
 
 namespace sqlpp
 {
   namespace mysql
   {
-    namespace detail
-    {
-      struct result_meta_data_t
-      {
-        size_t index;
-        unsigned long bound_len;
-        my_bool bound_is_null;
-        my_bool bound_error;
-        std::vector<char> bound_text_buffer;  // also for blobs
-        const char** text_buffer;
-        size_t* len;
-        bool* is_null;
-      };
-
-      struct prepared_statement_handle_t
-      {
-        struct wrapped_bool
-        {
-          my_bool value;
-
-          wrapped_bool() : value(false)
-          {
-          }
-          wrapped_bool(bool v) : value(v)
-          {
-          }
-          wrapped_bool(const wrapped_bool&) = default;
-          wrapped_bool(wrapped_bool&&) = default;
-          wrapped_bool& operator=(const wrapped_bool&) = default;
-          wrapped_bool& operator=(wrapped_bool&&) = default;
-          ~wrapped_bool() = default;
-        };
-
-        MYSQL_STMT* mysql_stmt;
-        std::vector<MYSQL_BIND> stmt_params;
-        std::vector<MYSQL_TIME> stmt_date_time_param_buffer;
-        std::vector<wrapped_bool> stmt_param_is_null;  // my_bool is bool after 8.0, and vector<bool> is bad
-        std::vector<MYSQL_BIND> result_params;
-        std::vector<result_meta_data_t> result_param_meta_data;
-        bool debug;
-
-        prepared_statement_handle_t(MYSQL_STMT* stmt, size_t no_of_parameters, size_t no_of_columns, bool debug_)
-            : mysql_stmt(stmt),
-              stmt_params(no_of_parameters, MYSQL_BIND{}),
-              stmt_date_time_param_buffer(no_of_parameters, MYSQL_TIME{}),
-              stmt_param_is_null(no_of_parameters, false),
-              result_params(no_of_columns, MYSQL_BIND{}),
-              result_param_meta_data(no_of_columns, result_meta_data_t{}),
-              debug(debug_)
-        {
-        }
-
-        prepared_statement_handle_t(const prepared_statement_handle_t&) = delete;
-        prepared_statement_handle_t(prepared_statement_handle_t&&) = default;
-        prepared_statement_handle_t& operator=(const prepared_statement_handle_t&) = delete;
-        prepared_statement_handle_t& operator=(prepared_statement_handle_t&&) = default;
-
-        ~prepared_statement_handle_t()
-        {
-          if (mysql_stmt)
-            mysql_stmt_close(mysql_stmt);
-        }
-
-        bool operator!() const
-        {
-          return !mysql_stmt;
-        }
-      };
-    }  // namespace detail
-
     class bind_result_t
     {
       std::shared_ptr<detail::prepared_statement_handle_t> _handle;
